HHH-13050 : Add test with a batch that fails when addToBatch() is called
(cherry picked from commit a7fccaa377b306b60102c868dc69b750c4d37ba8)
